# openapi/openapi.json
{
  "components": {
    "schemas": {
      "ActiveTaskItem": {
        "additionalProperties": false,
        "properties": {
          "age_s": {
            "title": "Age S",
            "type": "integer"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "entry_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "file_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "attempts",
          "age_s"
        ],
        "title": "ActiveTaskItem",
        "type": "object"
      },
      "ActiveTasksResponse": {
        "additionalProperties": false,
        "properties": {
          "pending": {
            "items": {
              "$ref": "#/components/schemas/ActiveTaskItem"
            },
            "title": "Pending",
            "type": "array"
          },
          "running": {
            "items": {
              "$ref": "#/components/schemas/ActiveTaskItem"
            },
            "title": "Running",
            "type": "array"
          }
        },
        "required": [
          "running",
          "pending"
        ],
        "title": "ActiveTasksResponse",
        "type": "object"
      },
      "BearerAuthError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "default": "missing or invalid bearer token",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "BearerAuthError",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecentTaskItem": {
        "additionalProperties": false,
        "properties": {
          "cache_creation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Creation"
          },
          "cache_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "entry_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "file_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "llm_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Calls"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "stages_ms": {
            "additionalProperties": true,
            "title": "Stages Ms",
            "type": "object"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tokens_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens In"
          },
          "tokens_out": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Out"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "label",
          "started_at",
          "finished_at",
          "last_error",
          "stages_ms"
        ],
        "title": "RecentTaskItem",
        "type": "object"
      },
      "RecentTasksResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecentTaskItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "RecentTasksResponse",
        "type": "object"
      },
...
      "RunningCountResponse": {
        "additionalProperties": false,
        "properties": {
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "running": {
            "title": "Running",
            "type": "integer"
          }
        },
        "required": [
          "running",
          "pending"
        ],
        "title": "RunningCountResponse",
        "type": "object"
      },
...
      "TaskThroughputCompleted": {
        "additionalProperties": false,
        "properties": {
          "done": {
            "title": "Done",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "files_per_minute": {
            "title": "Files Per Minute",
            "type": "number"
          },
          "success_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Rate"
          }
        },
        "required": [
          "done",
          "failed",
          "success_rate",
          "files_per_minute"
        ],
        "title": "TaskThroughputCompleted",
        "type": "object"
      },
      "TaskThroughputKindRow": {
        "additionalProperties": false,
        "properties": {
          "average_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Duration Seconds"
          },
          "completed_per_minute": {
            "title": "Completed Per Minute",
            "type": "number"
          },
          "done": {
            "title": "Done",
            "type": "integer"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "oldest_pending_age_seconds": {
            "title": "Oldest Pending Age Seconds",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "running": {
            "title": "Running",
            "type": "integer"
          },
          "success_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Rate"
          }
        },
        "required": [
          "kind",
          "pending",
          "running",
          "done",
          "failed",
          "oldest_pending_age_seconds",
          "average_duration_seconds",
          "success_rate",
          "completed_per_minute"
        ],
        "title": "TaskThroughputKindRow",
        "type": "object"
      },
      "TaskThroughputQueue": {
        "additionalProperties": false,
        "properties": {
          "oldest_pending_age_seconds": {
            "title": "Oldest Pending Age Seconds",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "running": {
            "title": "Running",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "pending",
          "running",
          "total",
          "oldest_pending_age_seconds"
        ],
        "title": "TaskThroughputQueue",
        "type": "object"
      },
      "TaskThroughputResponse": {
        "additionalProperties": false,
        "properties": {
          "by_kind": {
            "items": {
              "$ref": "#/components/schemas/TaskThroughputKindRow"
            },
            "title": "By Kind",
            "type": "array"
          },
          "completed": {
            "$ref": "#/components/schemas/TaskThroughputCompleted"
          },
          "queue": {
            "$ref": "#/components/schemas/TaskThroughputQueue"
          },
          "since": {
            "title": "Since",
            "type": "string"
          },
          "window_minutes": {
            "title": "Window Minutes",
            "type": "integer"
          }
        },
        "required": [
          "window_minutes",
          "since",
          "queue",
          "completed",
          "by_kind"
        ],
        "title": "TaskThroughputResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Library",
    "version": "0.3.6"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tasks/active": {
      "get": {
        "description": "Compact listing of running + pending tasks for the `/background` CLI.",
        "operationId": "list_active_v1_tasks_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveTasksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BearerAuthError"
                }
              }
            },
            "description": "Missing or invalid bearer token. Only raised when LIBRARY_API_TOKEN is configured. Tokenless loopback is the default.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Active",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/recent": {
      "get": {
        "description": "Recently-finished tasks (done + dead), newest first, with the\nper-run usage detail captured by the runner. Powers the StatusBar\nActivity popover so users can see how long ingest / reflect / embed\ntook and how many tokens each call burned.",
        "operationId": "list_recent_v1_tasks_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentTasksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BearerAuthError"
                }
              }
            },
            "description": "Missing or invalid bearer token. Only raised when LIBRARY_API_TOKEN is configured. Tokenless loopback is the default.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recent",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/running-count": {
      "get": {
        "description": "Count tasks currently in `running` or `pending` status.\n\nReturned counts include both states because in embedded mode,\npending tasks won't progress once the CLI exits either — the user\ncares about everything still on the queue, not just the in-flight\nrows.",
        "operationId": "running_count_v1_tasks_running_count_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunningCountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BearerAuthError"
                }
              }
            },
            "description": "Missing or invalid bearer token. Only raised when LIBRARY_API_TOKEN is configured. Tokenless loopback is the default.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Running Count",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/throughput": {
      "get": {
        "description": "Report ingest queue backlog and recent completion throughput.",
        "operationId": "task_throughput_v1_tasks_throughput_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_minutes",
            "required": false,
            "schema": {
              "default": 60,
              "title": "Window Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskThroughputResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BearerAuthError"
                }
              }
            },
            "description": "Missing or invalid bearer token. Only raised when LIBRARY_API_TOKEN is configured. Tokenless loopback is the default.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Throughput",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}